Extrace interfaces
