+++ core/src/main/java/com/linecorp/armeria/common/ContextAwareScheduledExecutorService.java
@@ -18,25 +18,19 @@ package com.linecorp.armeria.common;
 
 import static java.util.Objects.requireNonNull;
 
-import java.util.concurrent.Callable;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
-
-import com.google.common.base.MoreObjects;
 
 /**
  * A delegating {@link ScheduledExecutorService} that sets the {@link RequestContext} before executing any
  * submitted tasks.
  */
-public final class ContextAwareScheduledExecutorService
-        extends ContextAwareExecutorService implements ScheduledExecutorService {
+public interface ContextAwareScheduledExecutorService extends ScheduledExecutorService {
 
     /**
      * Returns a new {@link ContextAwareScheduledExecutorService} that sets the specified
      * {@link RequestContext} before executing any submitted tasks.
      */
-    public static ContextAwareScheduledExecutorService of(
+    static ContextAwareScheduledExecutorService of(
             RequestContext context, ScheduledExecutorService executor) {
         requireNonNull(context, "context");
         requireNonNull(executor, "executor");
@@ -49,62 +43,18 @@ public final class ContextAwareScheduledExecutorService
                     "cannot create a " + ContextAwareScheduledExecutorService.class.getSimpleName() +
                     " using another " + executor);
         }
-        return new ContextAwareScheduledExecutorService(context, executor);
-    }
-
-    private final ScheduledExecutorService executor;
-
-    private ContextAwareScheduledExecutorService(
-            RequestContext context, ScheduledExecutorService executor) {
-        super(context, executor);
-        this.executor = executor;
+        return new DefaultContextAwareScheduledExecutorService(context, executor);
     }
 
     /**
      * Returns the {@link ScheduledExecutorService} that is executing submitted tasks without setting
      * the {@link RequestContext}.
      */
-    public ScheduledExecutorService withoutContext() {
-        return executor;
-    }
+    ScheduledExecutorService withoutContext();
 
     /**
      * Returns the {@link RequestContext} that is specified when creating
      * this {@link ContextAwareScheduledExecutorService}.
      */
-    @Override
-    public RequestContext context() {
-        return super.context();
-    }
-
-    @Override
-    public ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit unit) {
-        return executor.schedule(context().makeContextAware(command), delay, unit);
-    }
-
-    @Override
-    public <V> ScheduledFuture<V> schedule(Callable<V> callable, long delay, TimeUnit unit) {
-        return executor.schedule(context().makeContextAware(callable), delay, unit);
-    }
-
-    @Override
-    public ScheduledFuture<?> scheduleAtFixedRate(Runnable command, long initialDelay, long period,
-                                                  TimeUnit unit) {
-        return executor.scheduleAtFixedRate(context().makeContextAware(command), initialDelay, period, unit);
-    }
-
-    @Override
-    public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command, long initialDelay, long delay,
-                                                     TimeUnit unit) {
-        return executor.scheduleWithFixedDelay(context().makeContextAware(command),
-                                                 initialDelay, delay, unit);
-    }
-
-    @Override
-    public String toString() {
-        return MoreObjects.toStringHelper(this)
-                          .add("executor", executor)
-                          .add("context", context())
-                          .toString();
-    }
+    RequestContext context();
 }
